Ref color
