--- main/srcs/base_object/color_debug.c
@@ -1,7 +1,0 @@
-#include "../../inc/minirt.h"
-
-int     print_color(t_color *color)
-{
-    printf("color: [%d, %d, %d]\n", color->r, color->g, color->b);
-    return (0);
-}
